+++ sdk/storage/Azure.Storage.Files.DataLake/src/DataLakePathClient.cs
@@ -11,6 +11,7 @@ using Azure.Core.Pipeline;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Specialized;
 using Azure.Storage.Files.DataLake.Models;
+using Azure.Storage.Sas;
 using Metadata = System.Collections.Generic.IDictionary<string, string>;
 
 namespace Azure.Storage.Files.DataLake
@@ -53,7 +54,7 @@ namespace Azure.Storage.Files.DataLake
         /// <summary>
         /// DFS Uri
         /// </summary>
-        internal Uri DfsUri => _dfsUri;
+        internal virtual Uri DfsUri => _dfsUri;
 
         /// <summary>
         /// Gets the directory's primary <see cref="Uri"/> endpoint.
@@ -377,7 +378,11 @@ namespace Azure.Storage.Files.DataLake
         /// The <see cref="ClientDiagnostics"/> instance used to create
         /// diagnostic scopes every request.
         /// </param>
-        internal DataLakePathClient(Uri pathUri, HttpPipeline pipeline, DataLakeClientOptions.ServiceVersion version, ClientDiagnostics clientDiagnostics)
+        internal DataLakePathClient(
+            Uri pathUri,
+            HttpPipeline pipeline,
+            DataLakeClientOptions.ServiceVersion version,
+            ClientDiagnostics clientDiagnostics)
         {
             var uriBuilder = new DataLakeUriBuilder(pathUri);
             _uri = pathUri;
@@ -386,7 +391,35 @@ namespace Azure.Storage.Files.DataLake
             _pipeline = pipeline;
             _version = version;
             _clientDiagnostics = clientDiagnostics;
-            _blockBlobClient = BlockBlobClientInternals.Create(_blobUri, _pipeline, Version.AsBlobsVersion(), _clientDiagnostics);
+            _blockBlobClient = BlockBlobClientInternals.Create(
+                _blobUri,
+                _pipeline,
+                Version.AsBlobsVersion(),
+                _clientDiagnostics);
+        }
+
+        internal DataLakePathClient(
+            Uri fileSystemUri,
+            string directoryOrFilePath,
+            HttpPipeline pipeline,
+            DataLakeClientOptions.ServiceVersion version,
+            ClientDiagnostics clientDiagnostics)
+        {
+            DataLakeUriBuilder uriBuilder = new DataLakeUriBuilder(fileSystemUri)
+            {
+                DirectoryOrFilePath = directoryOrFilePath
+            };
+            _uri = uriBuilder.ToUri();
+            _blobUri = uriBuilder.ToBlobUri();
+            _dfsUri = uriBuilder.ToDfsUri();
+            _pipeline = pipeline;
+            _version = version;
+            _clientDiagnostics = clientDiagnostics;
+            _blockBlobClient = BlockBlobClientInternals.Create(
+                _blobUri,
+                _pipeline,
+                Version.AsBlobsVersion(),
+                _clientDiagnostics);
         }
 
         /// <summary>
@@ -443,8 +476,8 @@ namespace Azure.Storage.Files.DataLake
                 var builder = new DataLakeUriBuilder(Uri);
                 _fileSystemName = builder.FileSystemName;
                 _accountName = builder.AccountName;
-                _path = builder.DirectoryOrFilePath;
-                _name = builder.LastDirectoryOrFileName;
+                _path = builder.DirectoryOrFilePath.UnescapePath();
+                _name = builder.LastDirectoryOrFileName.UnescapePath();
             }
         }
 
@@ -1565,12 +1598,25 @@ namespace Azure.Storage.Files.DataLake
                     }
 
                     // Build destination URI
-                    DataLakeUriBuilder destUriBuilder = new DataLakeUriBuilder(_dfsUri);
+                    DataLakeUriBuilder destUriBuilder = new DataLakeUriBuilder(_dfsUri)
+                    {
+                        Sas = null,
+                        Query = null
+                    };
                     destUriBuilder.FileSystemName = destinationFileSystem ?? destUriBuilder.FileSystemName;
-                    destUriBuilder.DirectoryOrFilePath = destinationPath;
 
-                    // We will get sas with destPath, if it was provided by the user.
-                    destUriBuilder.Sas = null;
+                    // DataLakeUriBuider will encode the DirectoryOrFilePath.  We don't want the query parameters,
+                    // especially SAS, to be encoded.
+                    string[] split = destinationPath.Split('?');
+                    if (split.Length == 2)
+                    {
+                        destUriBuilder.DirectoryOrFilePath = split[0];
+                        destUriBuilder.Query = split[1];
+                    }
+                    else
+                    {
+                        destUriBuilder.DirectoryOrFilePath = destinationPath;
+                    }
 
                     // Build destPathClient
                     DataLakePathClient destPathClient = new DataLakePathClient(destUriBuilder.ToUri(), Pipeline);