+++ sdk/storage/Azure.Storage.Files.DataLake/src/DataLakeUriBuilder.cs
@@ -110,6 +110,8 @@ namespace Azure.Storage.Files.DataLake
                 }
             }
         }
+
+        // Original path given by the user, unecoded.
         private string _directoryOrFilePath;
 
         /// <summary>
@@ -357,11 +359,12 @@ namespace Azure.Storage.Files.DataLake
                 {
                     if (DirectoryOrFilePath == "/")
                     {
-                        path.Append(DirectoryOrFilePath);
+                        path.Append(_directoryOrFilePath);
                     }
                     else
                     {
-                        path.Append("/").Append(DirectoryOrFilePath);
+                        // Encode path.
+                        path.Append("/").Append(_directoryOrFilePath.EscapePath());
                     }
                 }
             }