+++ sdk/storage/Azure.Storage.Files.Shares/src/ShareFileClient.cs
@@ -332,10 +332,10 @@ namespace Azure.Storage.Files.Shares
             if (_name == null || _shareName == null || _accountName == null || _path == null)
             {
                 var builder = new ShareUriBuilder(Uri);
-                _name = builder.LastDirectoryOrFileName;
+                _name = builder.LastDirectoryOrFileName.UnescapePath();
                 _shareName = builder.ShareName;
                 _accountName = builder.AccountName;
-                _path = builder.DirectoryOrFilePath;
+                _path = builder.DirectoryOrFilePath.UnescapePath();
             }
         }