+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 12.3.0-preview.1 (2020-06)
 - Added support for service version 2019-12-12.
 - Added support for Jumbo Files.
+- Fixed bug where DataLakeFileClient, DataLakeDirectoryClient, and DataLakePathClient.Name and .Path were sometimes URL-encoded.
 
 ## 12.2.2 (2020-06)
 - This release contains bug fixes to improve quality.